add instrument

+++ src/hook/PointerSymbolTable.h
@@ -13,9 +13,9 @@ struct PtrRecord {
   int16_t globalIdx;
   int16_t localIdx;
   int16_t action;
-  int64_t ptr;
-  int64_t size; // for alloca, and heap allocator
-  int64_t padding; // add something here?
+  uint64_t ptr;
+  uint64_t size; // for alloca, and heap allocator
+  uint64_t padding; // add something here?
 };
 
 constexpr size_t BUFFER_SIZE = 16*4096/sizeof(PtrRecord);
@@ -25,11 +25,11 @@ extern size_t K; // context length
 
 class PtaHook {
 private:
-  std::map<uintptr_t, std::pair<VId, size_t>> ptrToVid;
+  std::map<uint64_t, std::pair<VId, size_t>> ptrToVid;
   /// (Vid, K-context)
   std::unordered_map<VId, std::set<std::pair<VId, std::vector<VId>>>> pts;
   /// (function name, [alloca's pointers])
-  std::vector<std::pair<VId, std::vector<uintptr_t>>> scopeStack;
+  std::vector<std::pair<VId, std::vector<uint64_t>>> scopeStack;
 
 public:
   static void init(size_t k) { K = k; }
